+++ src/main/java/com/software/graduable/roadmap/RoadMapSemesterDTO.java
@@ -15,7 +15,7 @@ public class RoadMapSemesterDTO {
     // 과목명
     private String courseName;
     // 학점(설계)
-    private int credit;
+    private Double credit;
     //전공인지 교양인지
     private boolean category;
     private String yearAndSemester;
@@ -41,7 +41,7 @@ public class RoadMapSemesterDTO {
 
         return RoadMapSemesterDTO.builder()
                 .courseName(course.getCourseName())
-                .credit(course.getCredit().intValue())
+                .credit(course.getCredit())
                 .category(category)
                 .yearAndSemester(yearAndSemester)
                 .build();
